athene: perf: Decrease duration for app launch hint
Since the lineage qcom powerhal now sends POWER_HINT_LAUNCH when it
detects a long interaction, it now means that there are going to be more
times when the CPU cores woukd be pinned to the max speed for at least 3
seconds, which is detrimental to battery life. In addition, the
interactive governor ramps up the clock speed enough by itself, so
changing this wouldn't affect app launch speed by much.

Change-Id: I49cb66aadfc870f1a977bef40b2ea5ca244536f4

Signed-off-by: Akash Mondal <a98mondal@gmail.com>